Fixed target velocity

+++ ESP-Projects/main/CL_CAN.h
@@ -13,7 +13,7 @@
 //FUNCTIONS
 void talonPercentOut(int16_t speed);
 void setPIDValues();
-void setTargetVelocity(int16_t velocity);
+void setTargetVelocity(int velocity);
 void canSetup();
 
 #endif // CL_CAN_H
